Abstract to an adapter pattern.
