+++ adapters/fallback/lib/filterSort.js
