+++ adapters/fallback/lib/finder.js
