+++ adapters/fallback/lib/freshRequire.js
@@ -1,0 +1,7 @@
+module.exports = (path) => {
+  const key = Object.keys(require.cache).find((file) => {
+    return file.indexOf(path) !== -1
+  })
+  delete require.cache[key]
+  return require(path)
+}
