+++ adapters/fallback/lib/resolvepaths.js
