--- lib/freshRequire.js
@@ -1,4 +1,0 @@
-module.exports = (path) => {
-  delete require.cache[path]
-  return require(path)
-}
