+++ test/index.js
@@ -1,4 +1,4 @@
-require('./appFinder.spec')
-require('./fileFinder.spec')
-require('./appInfo.spec')
-require('./file.spec')
+require('./fallback/appFinder.spec')
+require('./fallback/fileFinder.spec')
+require('./fallback/appInfo.spec')
+require('./fallback/file.spec')
